+++ src/utils/join_all_split.c
@@ -1,0 +1,33 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   join_all_split.c                                   :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: tishihar <wingstonetone9.8@gmail.com>      +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2025/03/17 18:03:53 by tishihar          #+#    #+#             */
+/*   Updated: 2025/03/17 18:04:16 by tishihar         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "minishell.h"
+
+char	*join_all_split(char **array)
+{
+	char	*result;
+	char	*temp;
+
+	result = ft_strdup("");
+	if (!result)
+		return (NULL);
+	while (*array)
+	{
+		temp = result;
+		result = ft_strjoin(result, *array);
+		if (!result)
+			return (NULL);
+		free(temp);
+		array++;
+	}
+	return (result);
+}
